feat: start and stop single stack functions ready
