+++ utils/fn_stopStack.py
@@ -1,65 +1,53 @@
+import json
 import os
 import argparse
 import logging
-import time
-import docker
+import subprocess
+from colorama import Fore, Style, just_fix_windows_console
 from utils.cls import cls
-import json
+import time
 
-def stop_containers_from_file(containers_file: str):
-    client = docker.from_env()
-    try:
-        with open(containers_file, 'r') as f:
-            data = json.load(f)
-        containers = data.get('containers', {})
-        proxy_container = data.get('proxy', {})
+def stop_stack(compose_file: str) -> None:
+    """
+    Stop the Docker Compose stack using the provided compose file.
 
-        for app, container_list in containers.items():
-            for container_info in container_list:
-                container_name = container_info['container_name']
-                try:
-                    container = client.containers.get(container_name)
-                    container.stop()
-                    print(f"Stopped container: {container_name}")
-                except Exception as e:
-                    print(f"Failed to stop container {container_name}: {e}")
+    Arguments:
+    compose_file -- the path to the Docker Compose file
+    """
+    just_fix_windows_console()
 
-        if proxy_container:
+    while True:
+        response = input(f"{Fore.YELLOW}This will stop all the apps and delete the docker stack previously created using the configured docker-compose.yaml file. Do you wish to proceed (Y/N)?{Style.RESET_ALL} ").strip().lower()
+        if response in ['y', 'yes']:
             try:
-                container_name = proxy_container['container_name']
-                container = client.containers.get(container_name)
-                container.stop()
-                print(f"Stopped proxy container: {container_name}")
-            except Exception as e:
-                print(f"Failed to stop proxy container {container_name}: {e}")
+                result = subprocess.run(
+                    ["docker", "compose", "-f", compose_file, "down"],
+                    check=True,
+                    capture_output=True,
+                    text=True
+                )
+                print(f"{Fore.GREEN}All Apps stopped and stack deleted.{Style.RESET_ALL}")
+                logging.info(result.stdout)
+            except subprocess.CalledProcessError as e:
+                print(f"{Fore.RED}Error stopping and deleting Docker stack. Please check the configuration and try again.{Style.RESET_ALL}")
+                logging.error(e.stderr)
+            break
+        elif response in ['n', 'no']:
+            print(f"{Fore.BLUE}Docker stack removal canceled.{Style.RESET_ALL}")
+            time.sleep(1)
+            break
+        else:
+            print(f"{Fore.RED}Please answer yes or no.{Style.RESET_ALL}")
 
-        # Clear the containers file after stopping them
-        with open(containers_file, 'w') as f:
-            json.dump({'containers': {}}, f)
-    except Exception as e:
-        logging.error(f"An error occurred while stopping containers: {str(e)}")
-        raise
-
-def main(app_config_path: str, m4b_config_path: str, user_config_path: str, containers_file: str = './containers.json') -> None:
-    """
-    Main function to stop all running Docker containers.
-
-    Parameters:
-    app_config_path -- The path to the app configuration file.
-    m4b_config_path -- The path to the m4b configuration file.
-    user_config_path -- The path to the user configuration file.
-    containers_file -- The path to the containers JSON file.
-    """
-    cls()
-    stop_containers_from_file(containers_file)
-    logging.info("All Docker containers stopped successfully")
+def main(app_config_path: str, m4b_config_path: str, user_config_path: str) -> None:
+    compose_file = './docker-compose.yaml'
+    stop_stack(compose_file)
 
 if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Run the module standalone.')
+    parser = argparse.ArgumentParser(description='Stop the Docker Compose stack.')
     parser.add_argument('--app-config', type=str, required=False, help='Path to app_config JSON file')
     parser.add_argument('--m4b-config', type=str, required=True, help='Path to m4b_config JSON file')
     parser.add_argument('--user-config', type=str, required=False, help='Path to user_config JSON file')
-    parser.add_argument('--containers-file', type=str, default='./containers.json', help='Path to containers JSON file')
     parser.add_argument('--log-dir', default='./logs', help='Set the logging directory')
     parser.add_argument('--log-file', default='fn_stopStack.log', help='Set the logging file name')
     parser.add_argument('--log-level', choices=['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'], default='INFO', help='Set the logging level')
@@ -80,7 +68,7 @@ if __name__ == '__main__':
     logging.info("Starting fn_stopStack script...")
 
     try:
-        main(app_config_path=args.app_config, m4b_config_path=args.m4b_config, user_config_path=args.user_config, containers_file=args.containers_file)
+        main(app_config_path=args.app_config, m4b_config_path=args.m4b_config, user_config_path=args.user_config)
         logging.info("fn_stopStack script completed successfully")
     except FileNotFoundError as e:
         logging.error(f"File not found: {str(e)}")
